mshtml: Fix 'expando' and 'specified' for attributes in old compat modes.
'expando' returns TRUE on non-builtin attributes (builtin attributes are
only the builtin accessors, not methods), while 'specified' always returns
TRUE on any non-builtin attribute as long as it has a user-defined value
(e.g. for builtin methods, their slot must be non-default), which is an
extension to just checking the dynamic props.

Signed-off-by: Gabriel Ivăncescu <gabrielopcode@gmail.com>