light.msstyles: Introduce a CSS file to isolate colors from shapes.
The goal is to make it somewhat easier to define different color
schemes. It still won't be pretty since there is little tooling for MS
themes, but it will be nicer than having magic hex numbers in every SVG
file.

I don't know how standardized CSS in SVGs is. Imagick's convert
and Gimp support it in this way (<style>@import</style>). Another
suggestion I found is  to use <link xmlns="..." rel="stylesheet"
href="colors.css"/>, but I couldn't get it to work. KolourPaint supports
neither, so the highlight color will show up as black if you try to open
the SVG with it.