scrrun: drive_get_path implemented.
The function simply returns the drive letter and a colon (:)
for the passed drive. For example, if the passed drive is C,
it will return "C:".  These characters are found in the root
field of the drive and are the first two positions in the array.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=57539