winewayland: Ignore enter motion during pointer lock.
Having a pointer lock implies that relative motion is also used. When
refocusing on a mouselook application which the driver is using relative
motion for, an absolute motion hardware input is sent because of
handling the wl_pointer.enter event. This can result in an unwanted
warp/jerk.

The need to handle enter motion isn't applicable to mouselook or cases
where the pointer is locked because the application is drawing its own
cursor while covering vscreen, so it can be ignored during pointer lock.