winewayland: Ignore absolute motion during pointer lock.
This has the same idea as ignoring it if in relative mode, but the
effect is slightly different. Relative motion is disabled while pointer
focus is not on the constraint hwnd and re-enabled when pointer focus is
back on the constraint hwnd. An absolute motion event may be handled
before the foreground thread gets to re-enabling relative motion during
ClipCursor, which may result in an unwanted warp/jerk despite the
previous commit.