krnl386: Set carry flag for every failed DPMI call
According to the DPMI specification, when DPMI function fails the Carry
flag has to be set. Wine does not implement more DPMI function calls, so
it is needed to properly signal this fact to the called application.

For example if the application request for physical memory mapping via the
DPMI 0800h call (which is unsupported by both wine and Windows NT), it is
required to set carry flag to indicate failure. Otherwise it would be
treated as successful operation in which case in BX:CX registers is stored
linear address of the mapped memory to which application can access.
Obviously on successful path if the BX:CX registers are not set then
application would try to access random memory and crashes.

Signed-off-by: Pali Rohár <pali@kernel.org>