Implement ClassLibrary refcount leak test
Demonstrate a refcount leak on the class factory when a class with
ThreadingModel=Apartment is instantiated in the MTA. CoGetClassObject
ends up using apartment_hostobject to launch an STA thread in which to
host the object, but the process of marshaling this class factory back
into the MTA leaks a reference on it.

This leak is usually minor (class factory implementations are usually
singletons, so it's a one-off rather than a cumulative leak), but will
generally lock that COM server and e.g. prevent CoFreeUnusedLibraries
from being able to clean up properly. Since the refcount of an object
obtained from a hostapt in this fashion will generally be obscured by
by the resulting proxy/stub, test the leak via DllCanUnloadNow.

Signed-off-by: Kevin Puetz <PuetzKevinA@JohnDeere.com>