WIP: winewayland: Don't clip cursor coordinates to vscreen boundaries.
Since the positions of win32 windows and their corresponding Wayland
surfaces are not synchronized, there are cases where parts of a window
are outside the vscreen boundaries, and thus inaccessible to input
events, but still visible and accessible from a user (i.e., Wayland
compositor) standpoint. Try to remedy this issue by instructing the Wine
server to not clip the Wayland mouse event coordinates to vscreen
boundaries.