winewayland.drv: Treat configuration sequences as deltas
Configuration sequences completing with an xdg_surface.configure do not
necessarily contain all state. For example, for a toplevel, the
compositor might send xdg_surface.configure without sending an
xdg_toplevel.configure.

It is up to the client to maintain the sum of all previous configuration
sequences and to only modify those properties that are part of a
sequence.

This commit adds a new type wayland_surface_config_delta that contains a
wayland_surface_config but also a mask of all of those fields that have
been modified. The new function wayland_surface_config_apply_delta takes
care of transferring deltas downstream.