ntdll: Add stub ASan runtime.
This is a minimal runtime that gets wine to work with `--enable-sanitize`.

Explanation:

* ntdll: Add stubs for ASan runtime functions. They do nothing.
* ntdll: Minimal initialization for ASan. This includes mapping the shadow memory, which is accessed
  by compiler-inserted ASan instrumented; and setting value of required global variables inside
  ntdll.
    - Some ntdll functions that run before this initialization has to be annotated with
      `no_sanitize`. I moved `init_paths` to after this initialization, the other option is mark it
      `no_sanitize` as well.
    - Mapping shadow memory is done inside unix/virtual.c, because I have to create views for the
      shadow memory regions, to prevent ntdll from mapping over them later. I discovered this only
      happens some of the times and not sure why. I think this only happens with PROT_NONE mappings.

Fun facts:

The code to detect memory problems are not part of the ASan runtime (unless a special compile flag
is used). So with ASan enabled, the instrumented code will constant try to report problems, because
at this point the environment is not fully setup for ASan. The main problem being stack not being
unpoisoned when wine moves the stack pointer (think RtlUnwind, NtCallbackReturn, etc.). But
fortunately the function that reports error is part of the ASan runtime, which right now is a stub,
so nothing breaks. This is a good smoke test that the instrumentation hasn't terribly broken wine.