ntdll: Unpoison stack in __asan_handle_no_return.
ASan always assumes stack shadow is not poisoned upon function entry. This fact is guaranteed by the
fact that:

1. Uninstrumented code doesn't touch the shadow memory.
2. Stack shadow is always unpoisoned before a function returns.

However this assumption is broken by noreturn functions, functions that do not return like a normal
function would - think RtlUnwind, NtContinue, NtCallbackReturn. They might pop several function
frames off the stack, and none of them will have a chance to clear up their stack shadows. To solve
this problem, ASan insert a call to __asan_handle_no_return before a noreturn function is called.
This function is supposed to unpoison the entire stack (we will come back to this later). For wine,
this is a bit difficult, since wine has several different stacks: native PE, unix, and for the new
WoW64 mode, the 64-bit PE stack. And when __asan_handle_no_return is called on the unix side, it's
kind of difficult to find out where the stack pointer is on the PE side, and vice versa. To solve
that, we will exploit another ASan mechanism - the fake stacks. What it does isn't important right
now, but it allows us to hook into the prologue of every function. So if __asan_handle_no_return is
called on the unix side, we unpoison the unix stack, then set a flag that will cause the PE stack to
be unpoisoned the next time we enter a PE function's prologue, and vice versa. We also unpoison the
sigaltstack if __asan_handle_no_return is called from a signal handler.

To implement this, we need to store some per-thread state to keep track of the latest known stack
pointers on each of these stacks before a noreturn happened; as well as whether each stack needs to
be unpoisoned. Since we need this very early, before Tls is initialized, we have to smuggle this
data somewhere else. TEB->GdiTebBatch is already used to store wine private data
(struct ntdll_thread_data), so that seemed to be a natural place to put these. Some care was taken
to make sure the allocated memory can be accessed by a 32-bit process in the new WoW64 mode.

(Now, you might think clear the entire stack's shadow whenever a noreturn function is called is
wrong. And you would be right about that. Once the stack shadow is unpoisoned, we lose the ability
to detect stack memory misuses for the functions currently left on the stack, which will cause false
negatives. This is the same for vanilla ASan too, and should be mostly solved by implementing the
fake stacks.)