ntdll: Implement __asan_{un,}poison_stack_memory.
We needed an uninstrumented memset function for this. On PE side, the one in ntdll is instrumented,
which can't be used, so we add our own. We uses the no_builtin/inline-stringops attributes for clang
and gcc respectively to prevent them from being compiled to a call to memset. On the unix side, we
intend to add a memset wrapper (although there isn't one right now), so we use `dlsym` with
`RTLD_NEXT` to get the pointer to the real memset function. We can equally write our own memset here
as well.