ntdll: Implement ASan fake stack.
This is the fake stacks we have mentioned in a previous commit, which we said will reduce our false
negative rates. Its original purpose, is to catch stack use after returns, i.e. when you escape a
local variable. It does so by replacing the stack frame with a fake frame allocated from a stack
frame pool. When a function returns, its fake frame is not reused immediately, instead it's poisoned
and kept around for a bit, so access to it can be detected.

But how does it help with noreturn functions? First of all, fake frames are unpoisoned upon
allocation, so we never need to worry about unpoisoning them before a noreturn function. However,
fake frames are freed via a call to __asan_stack_free in function epilogues, when noreturn happens,
epilogues are skipped and we could leak fake frames. To solve that problem, we set a garbage
collection flag in __asan_handle_no_return, and later when we try to allocate a new frame, this flag
is checked and GC is performed when necessary. Expired fake frames are detected by comparing their
corresponding real stack pointers with the address of the current stack pointer. Which means we need
separate sets of fake frame pools for each real wine stack, to not confused the stack pointers.

Fake frame pools are sized based on the size of their corresponding real stack. A fake frames pool is
divided into 11 size classes, each size class contains as many bytes as the real stack. So the total
size of the pool is 11 times the size of the real stack. Kernel stack is a fixed 1MiB, so this will
be 11 MiB. User stack is dynamically sized, to avoid allocating too much memory, we cap it at 16 MiB
for 32-bit, or 256 MiB for 64-bit per size class, which means 176 MiB and 2.75 GiB for 32 and 64 bit
respectively.

Issues:

Several places in wine we assume that local variables are on the stack, which is not valid if we
are using fake stacks. Two such places are fixed in previous commits. Another one is in
`save_context` which is called by signal handlers, so we disable fake stack in signal handlers for
that. Hopefully we won't let to many problems slip through by doing so.