ntdll: Make sure to not write into poisoned memory in KeUserModeCallback.
In KeUserModeCallback we try to put a struct callback_stack_layout on to the PE stack. Depends on
how we entered kernel mode, there might be stale, poisoned stack frames just below the PE stack rsp.
This is because unpoisoning the stack can sometimes be deferred (see previous commits for more
info). In this case writing into a stack frame like this will trip ASan, so we unpoison the memory
we are about to write into first. Doing so is harmless because in this case, once we switch to PE,
it will immediately unpoison its stack anyway.