asan_dynamic_thunk: Add ASan dynamic thunk for DLLs.
The idea is to not export ASan symbols from ntdll.dll, generally speaking export new symbols from
ntdll.dll is a bad idea. And referencing ASan API functions through exported symbols from ntdll.dll
creates complex linking order problems.

Instead, we create a table of the pointers to the required functions, and store a pointer to this
table in the unused parts of the memory page used for the PEB (PEB64 is 0x7c8 bytes, less than a
page). This table is also duplicated for each DLL (by linking with libasan_dynamic_thunk.a), and is
initialized when the first ASan API function is called by the containing DLL. This is more inline
with how vanilla Windows ASan runtime works (minus the storing pointer in PEB part).

From this commit on, wine should be able to compile with ASan enabled.