winewayland.drv: Respond to size hints when moving out of fullscreen
Currently xdg_toplevel_config size hints are ignored if the config has state 0
to avoid spurrious re-sizing. However, if the window is currently configured
with a non-zero state, e.g. it is in fullscreen, then such an event from the
compositor indicates a switch to a floating window, in which case the size
hint should be respected.