mshtml: Fix node's onclick not responding.
When installing the photoshop-setup.exe version
of Photoshop 2025, there is no response after
clicking the continue button.

The reasons are as follows:
As a property of the node, setting events
such as onclick on the node does not
respond when clicked.

The following jscript code was executed
during the installation of Photoshop 2025:

$("#continueButton").attr("onclick",
  "continueBtnClicked('installOptionsContinueClicked');
  return false;");