urlmon/tests: Test output length calculation when output buffer is too small.
Native handles this poorly. When the buffer isn't large enough for URI
unescaping, native will return STATUS_SUCCESS with a 0 output length.

Besides that, on <= Windows 10 v1507, there are additional cases where the
returned length is 0. This appears to have been fixed in later versions.

Unsure if we need bug compatibility here.