win32u: Use the actual format when getting font metrics with GGO_METRICS.
Fix the rectangle from DrawText(..., DT_CALCRECT) with a selected font might not be enough to
actually draw text with the font for the EV Nova game.

This ensures that text extent calculated with a font selected in a device context is enough when
drawing texts using the same font and device context. The root cause is that FreeType might report
different font metrics for different GGO formats. GetTextExtentExPoint() gets text extent with
GGO_METRICS. And GGO_METRICS gets font metrics from glyphs loaded with FT_LOAD_TARGET_NORMAL, so
when drawing bitmap fonts, which is of GGO_BITMAP format loaded with FT_LOAD_TARGET_MONO, with the
calculated text extent, the result text might get truncated because of the font metrics difference.
On Windows, anti-aliased font and bitmap font of the same face font reports the same font metrics so
such a problem doesn't happen. The ultimate fix is to let FreeType hint the bitmap fonts with the
same font metrics as anti-aliased fonts but there are no settings that can achieve it unless we
write a new FreeType hinter.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=57309