services: Fix use-after-free during shutdown.
Since rpc processes requests on a work queue, it might hold a referecen to the
connection while the main thread is shutting down and freeing the database.
When the work item completes, it will release the connection which in turn
calls services.exe's rundown function which uses the freed database.