win32u: Limit the work area to the monitor rect.
When emulating a non-native aspect ratio display mode, calling map_monitor_rect() to map the
physical work area to the emulate work area might produce a rectangle larger than the emulated
display mode. For example, a physical work area (0,0,3840,2096) gets mapped to (0,-180,2560,1217)
when the emulated display mode is 2560x1080 and the physical display mode is 3840x2160. In such
cases, we should limit the work area to the emulated monitor rectangle.