ntdll: Detect the optical disc type on Linux using a SCSI command.
SCSI command 0x46 (GPCMD_GET_CONFIGURATION) returns in the seventh and
eighth bytes the "profile" that the disc drive is currently using, which
corresponds to the physical type of the disc in the drive. Profile
numbers 16 and up are used for DVDs and blu-rays, and Windows does not
distinguish between the two.

The GET CONFIGURATION request format, response format, and profile
numbers are defined on pages 1, 2, and 7 of
<https://www.t10.org/ftp/t10/document.97/97-263r0.pdf> and pages 66-71
of <https://www.t10.org/ftp/t10/document.05/05-206r0.pdf>.

More specific media information is still not implemented.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=57790